create test vote new signer to cw + migrate cw tests to community_wallet.rs
